+++ strucmotif-search-core/src/main/java/org/rcsb/strucmotif/domain/query/ContentType.java
@@ -6,7 +6,7 @@ import java.util.function.Predicate;
  * Controls the set of allowed targets, effectively providing high-level control to find exclusively PDB structures, or
  * exclusively computed structure models.
  */
-public enum StructureDeterminationMethodology implements Predicate<String> {
+public enum ContentType implements Predicate<String> {
     /**
      * Return only PDB-entries.
      */
@@ -14,12 +14,7 @@ public enum StructureDeterminationMethodology implements Predicate<String> {
     /**
      * Return only computed structure models.
      */
-    COMPUTATIONAL(s -> !s.matches(Constants.PDB_REGEX)),
-    // TODO impl INTEGRATIVE
-    /**
-     * Return all hits, regardless of provenance.
-     */
-    ALL(s -> true);
+    COMPUTATIONAL(s -> !s.matches(Constants.PDB_REGEX));
 
     private static class Constants {
         private static final String PDB_REGEX = "^[0-9][a-zA-Z0-9]{3}$";
@@ -27,7 +22,7 @@ public enum StructureDeterminationMethodology implements Predicate<String> {
 
     private final Predicate<String> condition;
 
-    StructureDeterminationMethodology(Predicate<String> condition) {
+    ContentType(Predicate<String> condition) {
         this.condition = condition;
     }